# Conflicts:
#	face/capture/src/main/java/com/simprints/face/capture/screens/livefeedback/LiveFeedbackFragmentViewModel.kt

+++ infra/config-store/src/main/java/com/simprints/infra/config/store/models/FaceConfiguration.kt
@@ -8,7 +8,7 @@ data class FaceConfiguration(
     val nbOfImagesToCapture: Int
         get() = rankOne?.nbOfImagesToCapture!!
 
-    val qualityThreshold: Int
+    val qualityThreshold: Float
         get() = rankOne?.qualityThreshold!!
 
     val imageSavingStrategy: ImageSavingStrategy
@@ -22,7 +22,7 @@ data class FaceConfiguration(
 
     data class FaceSdkConfiguration(
         val nbOfImagesToCapture: Int,
-        val qualityThreshold: Int,
+        val qualityThreshold: Float,
         val imageSavingStrategy: ImageSavingStrategy,
         val decisionPolicy: DecisionPolicy,
         val version: String,
